binary/writer: Remove all functionaly for Payload encoding
We no longer know how to write payloads, we simply write ArchiveHeader.
While this may seem a huge setback, the code for encoding was disgusting
and suffered many pattern failures.

We're going to use IOP abstraction and convert Payload into an actual
class, with an associated PayloadHeader struct encoded into the file.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>