binary/writer: Preliminary support for writing a Payload
We assemble a PayloadHeader manually from the given Payload, then
dump the PayloadHeader to the file. We do not allow the old seek
and rewrite approach from our first design, instead requiring that
all Payloads are assembled upfront, and then flushed.

Currently we do not actually encode the data *itself* as this is
implementation specific and not quite worked out yet.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>