binary: Add typed Payload registration system
Right now we don't *use* this system, but we will be using it to allow
automatic runtime construction of the appropriate Payload implementation
to handle decoding of the stream.

Vendors will be able to register their Payload to support automatic
decoding without modifying the codebase, allowing endless extensions
if so required.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>