binary: Encode the majority of the MetaPayload to stream
We're now able to iterate the in-memory pairs, and write the Record
struct for each one correctly to disk. We also calculate the CRC64ISO
as we go along for the decompressed data, as we only care for the
decompressed validity, not the compressed validity.

Next up is actually encoding the MetaPayload **values** and then
supporting ContentPayload.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>