db/encoding: Introduce the new db/encoding helper module
This module will power a large part of the contractual APIs for
automatic serialisation and deserialisation of `ubyte[]` into the
correct types requested during compilation, at runtime.

It is designed to be type-safe as well as extensive, so that we
can trivially implement new encode/decode routine to ensure things
are stored in an endian-independent fashion quite transparently.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>