query/manager: Rework manager to preload based APIs
We now have deduplicated preload APIs to load packages by name or ID
to ensure resolution is possible. Package IDs are the unique keys in
the runtime table so we disallow duplicates on this field.

We now use the queryProviders API and classify a Name as a provided
capability, allowing a preload by name into the runtime for further
manipulation.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>