analyser: Add our main control flow logic for one path
We'll hit the head of each chain until someone accepts (NextFunction) our
FileInfo. Control can be passed to the next chain with NextHandler or to
the next function in the chain with NextFunction.

The eventual desired outcome is to get an Ignore or Include order back
from the chains so we know whether to keep the file around or not, and
compute expensive hash sums only once.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>