registry/transaction: Add revdep removal support
It is not enough to just remove a package. Rather, we need to fully and
recursively remove anything depending on it otherwise we'll have a broken
dependency situation.

In order to achieve this we only consider the currently installed state
for any mutation or candidacy. A digraph is initiall built using these
rules and transposed so we can begin detection of reverse dependencies.

In order to isolate the dependencies we perform a topological sort of
the subgraph of each dependency transposition and then filter these
accumulated IDs from the final graph, before it is finally rebuilt in
the applicator step.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>