payload/index: Make our IndexEntry compact
We now drop a number of fields that aren't actually used anywhere:

 - size (this is now accessed by a runtime computation property)
 - refcount (can be trivially calculated at runtime)
 - length

The corresponding digest string that used to follow an IndexEntry has
been dropped in favour of merging a ubyte[16] field directly into the
IndexEntry. As such we now use all 32-bytes of the IndexEntry and
do not require a secondary NUL terminated string to be read in.

The impact on compression is negligible as its as compact now as it
can be, but at runtime we drop 33-bytes per IndexEntry (Pairing) so
our memory usage for the IndexPayload is roughly halved.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>