queue: Add a simple sorted dequeue
When fetching in a multithreaded environment we'd ideally dedicate
one thread for fetching the larger items, leaving other threads to
eat their way through all the smaller items.

To that purpose we have a multisorted set of fetchables with a way
to pop the head (smallest) or back (largest) elements so that some
download logic can be established.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>