logging: Redo our "should show" logic
We now only emit timestamps when the timestamps property is set.
We also will always emit file:line info when we get a critical
or fatal log message, as it's not relevant anywhere else.

Lastly we configure ColorLogger for *all* messages so that we can
accept them - and globalLogLevel will control the display of those
messages.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>