orm: Create and populate indices
An @Indexed field creates a reverse mapping of an owned datum,
i.e ".username" to permit quicker lookups when using an alternative
unique primary ID.

Note, we don't yet remove the *old* indexed data if it exists, so
we'll need to do that.

Another gotcha is this requires the use of publicly available
structs due to extensive CTFE/UDA use, but thats not *really*
a problem.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>