controller: Prefer large work items for most threads
Originally we thought we'd see better distribution by gobbling up the
small items on multiple threads but its a bad use of bandwidth from
what I'm seeing.

Worse than that, it gives the impression of downloads stalling towards
the end of the packages as we're waiting on the big ass packages to
download.

We now prefer large items first, with one thread dedicated to the small
items and working excellently with connection reuse, with the other threads
eating up the big packages to give the impression of a *speed up* towards
the end of the download stage.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>