registry/transaction: Mask broken fallthrough dependencies
In the interesting case of Perl, an older Perl provides libperl.so
and self-depends on it. In a newer build, it self-depends but does
not PROVIDE the symbol, so ends up depending on an old candidate
of itself which is just entirely screwed up.

To prevent these pass-through cases we mask out these broken-ass
dependency situations as a missing dependency from the low priority
situation, i.e the one not being actively selected that is attempting
to perform displacement.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>